requirements.txt
streamlit
openai
librosa
soundfile
yt-dlp
python-dotenv
anthropic
youtube_transcript_api
